Screen.flow
Added the 'myEV3' package to store useful classes and methods that can
be used anywhere.
Started with an (empty) Screen class to store the flow method that flows
a string on the display, using word wrap and supporting a scroll
mechanism. Words longer than 18 characters will simply flow onto the
next line.
Also created a simple program to test this method.