Implemented input methods
inputName() allows to enter a name on the EV3 screen at a position of
choice. it uses inputString()
inputString() allows text input using the EV3 keys: left or right to go
through the string, up or down to choose a letter from the alphabet.
-> Need to support inverted chracters

inputNumber() is a method to insert numbers using the EV3 keys. It
supports decimal numbers.