Optimised the flow
The main consist of startUp - selectBot - shutDown.
botNames is now a List and is only loaded at startup, or created if no
database existed.
New robots are put in the botNames list and stored in the roboBase file.
Previous selections are now kept in the menus.
A new robot name is compared to the existing set.
A new robot is created using default settings, and the settings are
shown for editing.
Any changes from the wheel calibration are applied.
The escape button now escapes properly to the previous screen.
TODO: escape in the wheel calibration